Uploaded Newer Version
In the original version I had the Spanish labels on skos:altLabel and the English labels on rdfs:label. I realized this is not good because it makes two different versions for English and Spanish. So I used a SPARQL query to:
1) Move the existing Spanish labels from skos:altLabel to rdfs:label
2) Set the language tag for the Spanish labels to @es (forgot to do that when I generated the labels the first time)
Now in Protege we can use File>Preferences>Renderer>Render by annotation property and in the configuration at the bottom of the Dialog there is an option to set the language tag. The default is en but if you change it to es then the Spanish labels display (I checked). So now this version can be used for both English and Spanish users.